feat: refactor components
