+++ daritza-teheran/tarea-7/src/components/Bottom.jsx
@@ -1,0 +1,10 @@
+import image from "../assets/codigo-de-barras.png";
+
+export default function Bottom({ data }) {
+  return (
+    <footer>
+      <p>{data}</p>
+      <img src={image} />
+    </footer>
+  );
+}
