+++ daritza-teheran/tarea-7/src/components/Lineimage.jsx
@@ -1,0 +1,12 @@
+import image from "../assets/image.jpg";
+
+export default function Linemage({ lines }) {
+  return (
+    <ul>
+      <img src={image} />
+      {lines.map((line) => (
+        <li key={line.toString()}>{line}</li>
+      ))}
+    </ul>
+  );
+}
