+++ daritza-teheran/tarea-7/src/components/Lines.jsx
@@ -1,0 +1,9 @@
+export default function Lines({ lines }) {
+  return (
+    <ul>
+      {lines.map((line) => (
+        <li key={line.toString()}>{line}</li>
+      ))}
+    </ul>
+  );
+}
