+++ daritza-teheran/tarea-7/src/data.js
@@ -1,0 +1,25 @@
+const masthead = {
+  title: "DOG WORLD",
+  subtitle: "All about your furry friend and more",
+};
+
+const lines = [
+  ["Caucasian Shepherd", "SPINONE", "Tibetan Mastiff"],
+  ["All about", "COCKER SPANIELS"],
+  ["+ Vet", "Lameness, a form of osteoarthritis"],
+];
+
+const lineImage = {
+  lines: ["+ Education", "What its behavior is so bad?"],
+};
+
+const bottom = {
+  data: "Dog world N°301 - mensual $3.5",
+};
+
+export { 
+    masthead,
+    lines,
+    bottom,
+    lineImage
+};
