+++ daritza-teheran/tarea-7/src/styles/components.css
@@ -1,3 +1,22 @@
+h1 {
+  color: white;
+  font-size: 10vh;
+  text-align: center;
+  margin: 0;
+  line-height: 0.65;
+  padding: 15px;
+}
+
+h2 {
+  color: #cc8334;
+  font-size: 2.5vh;
+  text-align: start;
+  margin: 0;
+  text-indent: 6.5vh;
+  font-weight: 500;
+  font-style: italic;
+}
+
 .cover-lines {
   display: flex;
   flex-direction: column;
