+++ daritza-teheran/tarea-7/src/styles/index.css
@@ -1,7 +1,9 @@
 @import url("https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100;0,400;0,500;1,100;1,300&display=swap");
+
 body {
   margin: 0;
 }
+
 * {
   font-family: "Montserrat", sans-serif;
 }
