--- daritza-teheran/tarea-7/src/styles/masthead.css
@@ -1,18 +1,0 @@
-h1 {
-  color: white;
-  font-size: 10vh;
-  text-align: center;
-  margin: 0;
-  line-height: 0.65;
-  padding: 15px;
-}
-
-h2 {
-  color: #cc8334;
-  font-size: 2.5vh;
-  text-align: start;
-  margin: 0;
-  text-indent: 6vh;
-  font-weight: 500;
-  font-style: italic;
-}
